Implement `status` subcommand and expand GitGlobalStatus.
Expand GitGlobalResult to include non-repo-specific messages and a flag for
inserting space between output for each repo. Update the JSON output structure
to reflect this.

For the list subcommand, push an empty string to each repo to force display of
the repo path.

For the scan subcommand (the impetus for the change to GitGlobalResult), add a
message indicating how many repos were found, and how to list them.

For the status subcommand, follow the example in the git2 source repo for
displaying a file's status in git's "short format". (Didn't implement the part
about submodules, though, because I don't use them personally. Added a TODO.)

This should just about complete the set of features planned for 0.1.0, but I'd
like to add some tests before calling that version final.