Move existing test to `tests/` directory.
The current test, and some other test case ideas that have the `ignore`
attribute right now, seem more like integration tests than unit tests,
so move them to `tests/`, as described in the new Rust book:
https://rust-lang.github.io/book/second-edition/ch11-03-test-organization.html

Also, move the repo-creation code to a `utils` module within the
`tests/` directory and add another utility function to create a set of
repos with a base directory. Add to-do for being able to inject
configuration for tests.