Add --untracked and --nountracked flags
These will override the show-untracked setting in ~/.gitconfig, if set,
in case you want to run the command once in a while with the opposite
setting.

I was hoping for some kind of boolean Arg from clap that would
automatically add a `--no` prefix variant, but I didn't find one.